HIVE-27186: final tests and code clean-up;
