+++ standalone-metastore/metastore-server/src/test/java/org/apache/hadoop/hive/metastore/properties/HMSThriftTest.java
@@ -23,14 +23,10 @@ import org.apache.hadoop.hive.metastore.MetaStoreTestUtils;
 import org.apache.hadoop.hive.metastore.conf.MetastoreConf;
 import org.apache.hadoop.hive.metastore.security.HadoopThriftAuthBridge;
 import org.apache.thrift.TException;
-import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Map;
-import java.util.TreeMap;
 
 public class HMSThriftTest extends HMSTestBase {
   /**
@@ -61,10 +57,6 @@ public class HMSThriftTest extends HMSTestBase {
         return null;
       }
     }
-//    @Override
-//    public Map<String, String> getProperties(List<String> keys) throws IOException {
-//
-//    }
   }
 
   @Override protected int createServer(Configuration conf) throws Exception {
